photon-order-created event
