+++ packages/components/src/systems/RecentOrders/RecentOrdersCombineDialog.tsx
@@ -14,6 +14,7 @@ import { dispatchDatadogAction } from '../../utils/dispatchDatadogAction';
 import { createMutation } from '../../utils/createMutation';
 import { Order } from '@photonhealth/sdk/dist/types';
 import { useDraftPrescriptions } from '../DraftPrescriptions';
+import { usePrescribeEventDispatch } from '../PrescribeEventDispatchProvider';
 
 const COMBINE_ORDERS_MUTATION = gql`
   mutation RecentOrdersCombineDialogUpdateOrder($orderId: ID!, $fills: [FillInput!]!) {
@@ -68,9 +69,8 @@ type VariablesCreateOrder = {
 
 export default function RecentOrdersCombineDialog() {
   let ref: Ref<any> | undefined;
-  const draftPrescriptionsContext = useDraftPrescriptions();
-
-  const { draftPrescriptions } = draftPrescriptionsContext;
+  const { draftPrescriptions } = useDraftPrescriptions();
+  const { dispatchOrderCreated } = usePrescribeEventDispatch();
 
   const client = usePhotonClient();
   const [state, actions] = useRecentOrders();
@@ -96,17 +96,6 @@ export default function RecentOrdersCombineDialog() {
       composed: true,
       bubbles: true,
       detail: { order }
-    });
-    ref?.dispatchEvent(event);
-  };
-
-  const dispatchOrderCreated = (order: SuccessResponse) => {
-    const event = new CustomEvent('photon-order-created', {
-      composed: true,
-      bubbles: true,
-      detail: {
-        order: order
-      }
     });
     ref?.dispatchEvent(event);
   };
@@ -170,7 +159,7 @@ export default function RecentOrdersCombineDialog() {
           }
         });
 
-        dispatchOrderCreated(newOrder.createOrder);
+        dispatchOrderCreated(newOrder.createOrder as Order);
         setIsCombiningOrders(false);
       } catch {
         triggerToast({
